+++ BE/src/main/java/refresh/ManageSystem/controller/annual/AnnualRestController.java
@@ -1,0 +1,2 @@
+package refresh.ManageSystem.controller.annual;public class AnnualRestController {
+}
