+++ BE/src/main/java/refresh/ManageSystem/controller/department/DepartmentRestController.java
@@ -1,0 +1,2 @@
+package refresh.ManageSystem.controller.department;public class DepartmentRestController {
+}
