--- BE/src/main/resources/static/src/pages/calendar/app.js
@@ -1,12 +1,0 @@
-import Calendar from './calendar.js';
-/*
-  Daniel Kim
-  
-  Calendar 함수 즉 dom을 리턴하는 함수
-  
-  2023-04-23
-*/
-
-export const $App = () => `
-  ${Calendar()}
-`
