+++ BE/src/main/resources/static/src/pages/calendar/aside/asideDepartment/app.js
@@ -1,0 +1,6 @@
+import AsideDepartment from './asideDepartment.js';
+
+
+export const $App = () => `
+  ${AsideDepartment()}
+`;
