+++ BE/src/main/resources/static/src/pages/calendar/aside/asideDepartment/main.js
@@ -1,0 +1,17 @@
+import { $App } from "./app.js";
+import { render } from "./hook.js";
+
+
+
+(function selectDepartment() {
+  const select = document.querySelector('#department-select');
+  select.onchange = (e) => {
+    const department = e.target.value;
+    render($App, document.querySelector("#aside"), department);
+  }
+})();
+
+
+
+
+
