+++ BE/src/main/resources/static/src/pages/calendar/aside/asideHeader/app.js
@@ -1,0 +1,6 @@
+import AsideHeader from './asideHeader.js';
+
+
+export const $App = () => `
+  ${AsideHeader()}
+`;
