+++ BE/src/main/resources/static/src/pages/calendar/aside/asideMember/app.js
@@ -1,0 +1,6 @@
+import AsideMember from './asideMember.js';
+
+
+export const $App = () => `
+  ${AsideMember()}
+`;
